Mainstream Framework Integration/scipy_agent/requirements.txt
aspose-cells-python
numpy
pandas
scipy
scikit-learn
statsmodels
openpyxl
matplotlib
openai
python-dotenv
pydantic